### 2020-05-24 #2248 @whedon
[PRE REVIEW]: Excimontec v1.0: An Open-Source Software Tool for Kinetic Monte Carlo Simulations of Organic Electronic Devices
**Submitting author:** @MikeHeiber (<a href="http://orcid.org/0000-0002-1567-5663">Michael Heiber</a>)
**Repository:** <a href="https://github.com/MikeHeiber/Excimontec" target ="_blank">https://github.com/MikeHeiber/Excimontec</a>
**Version:** v1.0.0
