### 2020-05-24 #2248 @arfon
Copying an excerpt of the author submission comment for future editors which includes potential reviewers:

> I'd like to suggest several expert reviewers that may not already be current JOSS reviewers: Prof. Geoff Hutchison, University of Pittsburgh, `@ghutchis` on GitHub. Dr. Christoph Junghans, Los Alamos National Lab,`@junghans` on GitHub Prof. Bjoern Baumeier, Eindhoven University of Technology, `@baumeier` on GitHub
