### 2020-06-04 04873fb
Refactor Update
- Updated copyright for all files changed in this PR
- Removed testing files energies_missing_data and energies_missing_dims because these are now generated automatically during the testing process
- Added parameters_v1.0.txt to the test dir to be used to test parameter file format from v1.0

parameters_default.txt
- Updated parameter list include State_saving_interval parameter that determines how often the state is saved during the equilibration phase of the steady transport test
- Fixed issue with how the site energy file is designated for importing site energies

main.cpp
-
[…]
