+++ src/Parameters.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2019 Michael C. Heiber
+// Copyright (c) 2017-2020 Michael C. Heiber and contributors
 // This source file is part of the Excimontec project, which is subject to the MIT License.
 // For more information, see the LICENSE file that accompanies this software.
 // The Excimontec project can be found on Github at https://github.com/MikeHeiber/Excimontec
@@ -139,6 +139,10 @@ namespace Excimontec {
 		}
 		if (Enable_steady_transport_test && Enable_bilayer) {
 			cout << "Error! When running a steady transport test, the bilayer device architecture cannot be used." << endl;
+			return false;
+		}
+		if (State_saving_interval < 10) {
+			cout << "Error! Steady transport test state saving interval must not be less than 10." << endl;
 			return false;
 		}
 		// Check exciton parameters
@@ -299,15 +303,10 @@ namespace Excimontec {
 			cout << "Error! When importing site energies from a file, the interfacial energy shift model must not be enabled." << endl;
 			return false;
 		}
-		if (Enable_import_energies && (int)Energies_import_filename.size() == 0) {
-			cout << "Error! When importing site energies from a file, a valid filename must be provided." << endl;
+		if (Enable_import_energies && (int)Energies_import_format.size() == 0) {
+			cout << "Error! When importing site energies from a file, a valid file naming format must be provided." << endl;
 			return false;
 		}
-        if (Output_interval < 1) {
-            cout << "Error! Output interval is smaller than 1." << endl;
-            return false;
-        }
-        
 		// Check Coulomb interaction parameters
 		if (!(Coulomb_cutoff > 0)) {
 			cout << "Error! The Coulomb cutoff radius must be greater than zero." << endl;
@@ -320,17 +319,33 @@ namespace Excimontec {
 		return true;
 	}
 
-	bool Parameters::importParameters(ifstream& inputfile) {
+	bool Parameters::importParameters() {
+		// open the parameter file
+		ifstream parameterfile;
+		parameterfile.open(Parameters_filename, ifstream::in);
+		if (!parameterfile.good()) {
+			cout << "Error loading parameter file.  Program will now exit." << endl;
+			return false;
+		}
+		// load and parse the lines of the file
 		string line;
 		string var;
 		size_t pos;
 		vector<string> stringvars;
 		bool Error_found = false;
-		if (!inputfile.is_open() || !inputfile) {
+		if (!parameterfile.is_open() || !parameterfile) {
 			throw invalid_argument("Error importing parameter file because ifstream cannot read the parameter file.");
 		}
-		while (inputfile.good()) {
-			getline(inputfile, line);
+		// get paramterfile version from first line comment
+		Version params_version;
+		if (parameterfile.good()) {
+			getline(parameterfile, line);
+			pos = line.find_first_of("v", 0);
+			params_version = Version(line.substr(pos + 1));
+			cout << params_version.getVersionStr();
+		}
+		while (parameterfile.good()) {
+			getline(parameterfile, line);
 			if ((line.substr(0, 2)).compare("--") != 0 && (line.substr(0, 2)).compare("##") != 0 && line.compare("") != 0) {
 				pos = line.find_first_of("/", 0);
 				var = line.substr(0, pos);
@@ -338,6 +353,9 @@ namespace Excimontec {
 				stringvars.push_back(var);
 			}
 		}
+		// close parameter file
+		parameterfile.close();
+		// Assign the line values to the appropriate variables
 		int i = 0;
 		// KMC Algorithm Parameters
 		try {
[…]
