+++ src/main.cpp
@@ -1,42 +1,34 @@
-// Copyright (c) 2017-2019 Michael C. Heiber
+// Copyright (c) 2017-2020 Michael C. Heiber and contributors
 // This source file is part of the Excimontec project, which is subject to the MIT License.
 // For more information, see the LICENSE file that accompanies this software.
 // The Excimontec project can be found on Github at https://github.com/MikeHeiber/Excimontec
 
 #include "OSC_Sim.h"
 #include "Parameters.h"
-#include <mpi.h>
+#include <ctime>
 #include <fstream>
+#include <functional>
 #include <iostream>
-#include <cstdio>
-#include <experimental/filesystem>
+#include <mpi.h>
 #include <string>
 #include <vector>
-#include <ctime>
-#include <functional>
 
 using namespace std;
 using namespace Excimontec;
 using namespace KMC_Lattice;
 
 int main(int argc, char *argv[]) {
-	string version = "v1.0.0-rc.3";
+	string version = "v1.1.0-beta.1";
 	// Parameters
 	bool End_sim = false;
 	// File declaration
-	ifstream parameterfile;
 	ofstream logfile;
 	ofstream resultsfile;
 	ofstream analysisfile;
 	// Initialize variables
 	string logfilename;
-	string energies;
-	string occupance_fname;
-    string old_occup_fname;
 	Parameters params;
 	int nproc = 1;
-	int procid = 0;
-    int rand_seed = 0;
 	int elapsedtime;
 	time_t time_start, time_end;
 	bool success;
@@ -45,71 +37,33 @@ int main(int argc, char *argv[]) {
 	vector<bool> error_status_vec;
 	vector<string> error_messages;
 	char error_found = (char)0;
-    // count nr of command line parameters above the standard two
-    int accounted_cmd_pars = 2;
 	// Start timer
 	time_start = time(NULL);
-	// Check command line arguments
-	if (argc < 2) {
-		cout << "Error! You must input the parameter file name as a command line argument." << endl;
+	// Initialize mpi options
+	cout << "Initializing MPI options... ";
+	MPI_Init(&argc, &argv);
+	MPI_Comm_size(MPI_COMM_WORLD, &nproc);
+	MPI_Comm_rank(MPI_COMM_WORLD, &params.Proc_ID);
+	cout << params.Proc_ID << ": MPI initialization complete!" << endl;
+	// Initialize error monitoring vectors
+	proc_finished.assign(nproc, false);
+	error_status_vec.assign(nproc, false);
+	error_messages.assign(nproc, "");
+	// Parse command line arguments
+	success = params.parseCommandLineArguments(argc, argv);
+	if (!success) {
 		return 0;
 	}
-	// Check for command line enabled logging
-	// Set default
-
-    string argument;
-	params.Enable_logging = false;
-	if (argc > 2) {
-        for (int idx=2; idx < argc; idx++) {
-            argument = argv[idx];
-
-            if (argument.compare("-enable_logging") == 0) {
-                params.Enable_logging = true;
-                accounted_cmd_pars += 1;
-            }
-            if (argument.compare("-random_seed") == 0) {
-                if (argc >= idx+1) {
-                    rand_seed = stoi(argv[idx+1]);
-                    accounted_cmd_pars += 2;
-                    idx++;
-                }
-                else {
-                    cout << "Error! invalid random seed given." << endl;
-		            return 0;
-                }
-		    }
-        }
-	}
-    
-	// Check for too many command line arguments
-	if (argc != accounted_cmd_pars) {
[…]

[…]
