### 2020-06-08 #2248 @pdebuyl
@elena-pascal , @dalonsoa are you willing to review this paper for JOSS ?
