### 2020-06-08 #2248 @pdebuyl
@mdavezac would you review the article "Excimontec v1.0: An Open-Source Software Tool for Kinetic Monte Carlo Simulations of Organic Electronic Devices" for the Journal of Open Source Software?
