### 2020-06-08 #2248 @elena-pascal
@pdebuyl sorry my C++ is not good enough to review this.
