### 2020-06-08 #2307 @whedon
[REVIEW]: Excimontec v1.0: An Open-Source Software Tool for Kinetic Monte Carlo Simulations of Organic Electronic Devices
**Submitting author:** @MikeHeiber (<a href="http://orcid.org/0000-0002-1567-5663">Michael Heiber</a>)
**Repository:** <a href="https://github.com/MikeHeiber/Excimontec" target ="_blank">https://github.com/MikeHeiber/Excimontec</a>
**Version:** v1.0.0

@mdavezac, @Luthaf, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
