### 2020-06-16 #2307 @pdebuyl
@arfon is it possible for me to restart an invitation to the reviewer group?
