### 2020-06-16 #2307 @arfon
It is, you can do `@whedon re-invite @reviewer as reviewer`
