### 2020-06-24 #2307 @mdavezac
## Compilation

The procedure completed almost without a hitch. The issue is that it requires the users to set up their environment with a CXX variable. This may be a standard for  a developer, not so much for user. This requirement is missing from the installation guide.

What's weird though is that the environment variable is not actually used to identify the compiler, since the compiler is **hard-coded** to mpicxx.  It's merely a surprising way to specify a set of CXXFLAGS. However, even though CXX is just a switch, it cannot be set in the expected way for  a makefile, e.g. `make CXX=xxx`. It has to be done via an environment variable. And the environment variable cannot be something like `g++-8.0`. Hence, the setup will fail in any number of standard situations where CXX points to an actual versioned executable.

Among other environments, this makefile will fail on crays, where mpi compiler wrappers have less standard names.

## Installation

There is no installation procedure per se. However, the main product is a relocatable executable.

## Recommendation

At the low end, it would be sufficient to explain the non-standard usage of CXX in the user-guide and that the executable can be relocated to another location of the user's choosing. It may be  good idea to replace CXX with a suitably named makefile variable with a suitable default.

However, I would **strongly** recommend ditching hand-rolled makefiles in favor of [meson](https://mesonbuild.com) or [cmake](https://www.cmake.org). The advantages will include: (i) out of tree builds, (ii) large set of compilers, (iii) standard installation procedure to /usr/bin, (v) installation of libKMC.a (or .so),  (v) build specification shorter and less error prone than a makefile.
Furthermore, it would give the author the opportunity to replace the git submodules with either [meson's dependency wrapper](https://mesonbuild.com/Wrap-dependency-system-manual.html) or [cmake's external project](https://cmake.org/cmake/help/latest/module/ExternalProject.html). Submodules are a source of confusion even for expert git users, especially in a multi-developer environment.
