### 2020-06-24 #2307 @mdavezac
Software paper: Travis provide continuous integration. [Integration testing](https://en.wikipedia.org/wiki/Integration_testing) is something else still. Indeed some of the unit-tests in the code could be termed integration tests, e.g. ExcitonDiffusionTests.
