### 2020-06-24 #2307 @pdebuyl
Thank you @mdavezac for the review!

@MikeHeiber you can already reply to @mdavezac , you don't have to wait for @liangtang 's review.

The process at JOSS is to discuss the comments as a dialogue on this github issue page.
