### 2020-06-30 #2307 @pdebuyl
@amandadumi would you review the article "Excimontec v1.0: An Open-Source Software Tool for Kinetic Monte Carlo Simulations of Organic Electronic Devices" for the Journal of Open Source Software?
