### 2020-07-02 #2307 @pdebuyl
Thank you @Luthaf for accpeting the invitation.

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know.
