### 2020-07-02 #2307 @danielskatz
note that command did not do what you intended
you now have 2 reviewers
1 - @mdavezac
2 - @mdavezac @Luthaf
whedon only can handle one at a time.  You should remove "@mdavezac @Luthaf" and then add "@Luthaf"
