### 2020-07-02 #2307 @Luthaf
I accepted the invite, but I can't check the boxes in the initial message. Any idea on what might be wrong?
