### 2020-07-02 #2307 @Luthaf
I reloaded the page in the mean time, and it worked (hence I deleted the comment visible in @pdebuyl email answer). Everything is fine now!
