### 2020-07-06 MikeHeiber/Excimontec#107 @Luthaf
Missing FLAG errror when building with clang
This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/2307

I tried to build Excimontec using clang as the C++ compiler, which failed to compile since there is no `FLAGS` defined in the makefile for this compiler. Strangely, the condition trying to catch this

```makefile
ifndef FLAGS
	$(error Valid compiler not detected.)
endif
```

did not catch anything, and instead the compilation failed due to C++11 specific constructs without `-std=c++11`. It looks like this code is not serving its intended purpose as far as I can understand it.

----

As a side note, using the same flags as GCC with clang makes the code compile fine and the tests pass on macOS, so I think support for clang is only a matter of detecting it.
