### 2020-07-06 #2307 @Luthaf
I agree with @mdavezac comment regarding compilation (https://github.com/openjournals/joss-reviews/issues/2307#issuecomment-648739993); and also recommend using CMake/Autotools/Meson to build Excimontec (my personal preference being for CMake). I don't mind the use of git submodules, if they are well integrated with the build system everything should be fine.

Since changing the build system can be a bit of work (although very much worth it); you could at least fix the issue around CXX by using `mpicxx -showme` to get the actual compiler name, and potentially running `$(compiler) --version` to detect gcc/clang/pgi/cray compilers/...
