### 2020-07-09 #2307 @MikeHeiber
Thank you @mdavezac and @Luthaf for the feedback about the build system Excimontec is currently using.  While it might not be the most ideal solution, building the tool with GNU Make is fairly simple and it works.  I've compiled and run Excimontec using Make on a number of different computing clusters at different universities and national labs, and I've never had major issues. There's also a handful of current users around the world using the tool, and they also have not reported issues with the current build system. For this reason, it is not a high priority to update to a more sophisticated build system at this time. There is an issue filed to switch to cmake (https://github.com/MikeHeiber/Excimontec/issues/25), and I'll be monitoring users' interest in this feature as time goes on. Currently, there's mostly only users and only myself as the main developer, so feature prioritization is really critical. I tend to favor working on features that improve functionality whenever possible to continue enabling the cutting edge research being done by users.

The way the compilation works right now is not a strange usage of the CXX environment variable.  GNU Make sets the CXX variable based on the compiler that it is set up to use. I have never needed to manually set the CXX variable before compilation. The makefile is simply detecting the compiler type that the user has set up on their system to set the compiler flags appropriately. This might not work if the user has a bunch of different compiler versions set up on their system and wants to specifically use a particular version, but in this more advanced case the super user that is interested in doing this can also simply modify the makefile to conform to their system. However, I will look into @Luthaf 's suggestions about potentially better ways to determine the compiler being used under the hood of the mpicxx command and adding support for the clang compiler.
