### 2020-07-14 #2307 @mdavezac
@MikeHeiber, I'm not sure what to say. Here are two random reviewers and both found that the build did not go without a hitch... You may be overestimating the time it would take to convert to a modern build system. For a project this size and with so few dependencies, it would not be much more time then re-inventing compiler/mpi detection in GNU Make. In any case, are you planning to **investigate this for this submission**? Otherwise, I would recommend explaining in the manual how CXX is used, or at least list the values it can take if set.

Also, I would appreciate if you could take this comment into account: https://github.com/openjournals/joss-reviews/issues/2307#issuecomment-648748916
