### 2020-07-14 MikeHeiber/Excimontec#108 @Luthaf
Missing dependencies in installation instruction
The user manual seems to assume that GNU Make and GCC are already installed in Linux installation instruction. While this should hold true for most super-computers, it might not be the case in new installations/local machine.

You may want to state that GCC and GNU Make are required to build Excimontec; or even better add them to the `apt-get` command page 19 of the user manual.

Ref https://github.com/openjournals/joss-reviews/issues/2307
