### 2020-07-15 #2307 @pdebuyl
Hi all,

At JOSS, we aim at improving the quality of the submitted software. However, mandating a specific build system (or otherwise specific features) is not part of our review. Besides, GNU Make is widely used and supported.

What is part of the review criteria is that the installation instructions are functional and properly documented, and that is what the review should focus on as far as the build system is concerned.
