+++ .travis.yml
@@ -5,36 +5,79 @@
 # Travis CI continuous integration build script
 
 language: cpp
-os: linux
-dist: trusty
+sudo: required
 
-compiler:
-    - gcc
-    
-env:
-    - OPTIONS="GCC_VERSION=4.9 && MPI=openmpi && TARGET=test_coverage"
-    - OPTIONS="GCC_VERSION=4.9 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=4.9 && MPI=openmpi && TARGET=test"
-    - OPTIONS="GCC_VERSION=5 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=5 && MPI=openmpi && TARGET=test"
-    - OPTIONS="GCC_VERSION=6 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=6 && MPI=openmpi && TARGET=test"
-    - OPTIONS="GCC_VERSION=7 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=7 && MPI=openmpi && TARGET=test"
-    - OPTIONS="GCC_VERSION=8 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=8 && MPI=openmpi && TARGET=test"
-    - OPTIONS="GCC_VERSION=9 && MPI=mpich2 && TARGET=test"
-    - OPTIONS="GCC_VERSION=9 && MPI=openmpi && TARGET=test"
-         
+matrix:
+  include:
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=5.4.0 && MPI=openmpi && TARGET=test_coverage"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=5 && MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=6 && MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=7 && MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=8 && MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=9 && MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=5.4.0 && MPI=mpich && TARGET=test"
+    - os: linux
+      dist: bionic
+      compiler: gcc
+      env:
+        - OPTIONS="GCC_VERSION=7.4.0 && MPI=mpich && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: clang
+      env:
+        - OPTIONS="MPI=openmpi && TARGET=test"
+    - os: linux
+      dist: xenial
+      compiler: clang
+      env:
+        - OPTIONS="MPI=mpich && TARGET=test"
+
 before_install:
     - eval "${OPTIONS}"
-    - sudo add-apt-repository ppa:ubuntu-toolchain-r/test -y
-    - sudo apt-get update -qq
-    - sudo apt-get install -qq gcc-$GCC_VERSION g++-$GCC_VERSION -y
[…]
