### 2020-07-24 #2307 @MikeHeiber
To improve the build process, I've updated how the makefile detects the compiler that the user has installed to run with their MPI installation.  Similar to what @Luthaf suggest, the default makefile now uses the output of the `mpicxx -show` command to determine the compiler that is installed and sets the compiler flags accordingly.  This should be more robust than using the CXX environment variable as was being done previously.  I've also added clang support by adding it to the makefile compiler detection list and also by including it in my TravisCI build matrix. The latest version can be seen here:
https://github.com/MikeHeiber/Excimontec/tree/joss-paper-update

While I am working on the other improvements, such as some additions to the User Manual, please finish the other parts of the review. There should hopefully be nothing with the build process that is blocking further review progress.
