### 2020-07-26 f726f95
Makefile and build testing update
Updated KMC_Lattice submodule to lastest version
.travis.yml - Simplified how GCC versions are installed and configured
.travis.yml - Clarified when default versions of GCC are used in the build matrix
makefile - Added detection of versioned g++ compilers
makefile - Removed unnecessary installation of clang during clang build testing
