### 2020-07-26 #2307 @MikeHeiber
I've made another update based on a suggestion by @Luthaf and the makefile should now be able to handle situations where the compiler underlying mpicxx is a versioned g++ executable.  For example, g++-9 will still be recognized as a g++ family compiler and the compiler flags will be set accordingly.
