### 2020-07-29 MikeHeiber/Excimontec#107 @Luthaf
I still have an error while building the code on the `joss-paper-update` branch because KMC_Lattice Makefile is not updated to the new compiler detection. If I copy/paste the corresponding code in the `KMC_Lattice/makefile` it works fine.
