### 2020-07-29 #2307 @Luthaf
The build is still failing for me with clang compilers: https://github.com/MikeHeiber/Excimontec/issues/107#issuecomment-665568938

I also found a problem with character encoding in result files https://github.com/MikeHeiber/Excimontec/issues/110, although I do not consider it to be a blocking point for this review.

---

Regarding the paper itself, I am unable to comment on

> **State of the field**: Do the authors describe how this software compares to other commonly-used packages?

since I've never used or heard of KMC simulation before.

The paper is fairly short, and building on top two other JOSS papers by the same author: https://joss.theoj.org/papers/10.21105/joss.01072 and https://joss.theoj.org/papers/10.21105/joss.01168. It might benefit from a short description of the core KMC simulation idea; as well as how it is applied to organic electronic devices.
