### 2020-08-03 MikeHeiber/Excimontec#107 @MikeHeiber
@Luthaf The current version on the `joss-paper-update` branch should work.  It's currently working with clang++ on Ubuntu in the TravisCI tests.  Make sure you are updating the submodule using `git checkout --recurse-submodules`.
