### 2020-08-09 f9fae39
Paper and User Manual Update
[skip ci]
paper.md
- Added additional information about what KMC simulations can be used for and specifically what types of simulation tests come with v1.0
- Add specification that TravisCI is doing continuous integration tests instead of simply calling them integration tests
User manual:
- Updated manual to show that parameters file section in each section of the manual and explain more directly what the parameters are used for
- Added information about the parameter file format and syntax
- Added installation instructions for GCC compiler and GNU Make
