+++ user_manual/main.tex
@@ -46,7 +46,7 @@ The software is written in modern C++ and is optimized for efficient pleasingly 
 This software package uses object-oriented design and extends the KMC\_Lattice framework.\cite{heiber2019joss} 
 It is also designed to work with the Ising\_OPV software tool for creating and importing bulk heterojunction morphologies.\cite{heiber2018joss}
 
-\section{Simulation Tests}
+\section{Simulation Tests and Their Parameters}
 
 Excimontec is designed to allow users to select and run a particular simulation test with a specific set of input parameters.
 Currently there are five different simulations tests available:
@@ -57,13 +57,52 @@ Currently there are five different simulations tests available:
     \item Dynamics Test
     \item Steady State Charge Transport Test
 \end{enumerate}
-Choice of a particular test and the input parameters is made within the parameter file that is then passed as an input argument to the program.
+Choosing a particular test and setting the input parameters is made within the parameter file that is then passed as an input argument to the program.
 An example parameter file (parameters\_default.txt) is provided in the root repository directory to be used as a starting point for customization.
-A valid parameter file has strict formatting rules, so users should simply copy the default parameter file and then modify the values as needed.
+
+A single parameter file is used to run the KMC simulation test for a single set of parameters.
+To do parameter sweeps, one must create separate parameter files for each condition and then run separate simulations for each.
+While this can be tedious when manually creating and modifying the parameter files, shell scripts can be used to quickly create a series of parameter files for a parameter sweep, and simulation job arrays can be used to submit all of the simulations in a parameter sweep and group the individual simulations together.
+
+The parameter file has a fixed format and the program parses the parameter values in a set order to define the internal variables specified in the comments that follow each value.
+The comments starting with "//" that trail each parameter definition are ignored by the parser and are simply provided to inform the user of what each line represents and give information about the expected data type or other restrictions.
+Moving variable definitions to different positions in the file will cause an error or incorrect assignment of parameter values.
+Parameter sections headings that begin with "\#\#" and hyphen line breaks are ignored by the parser. 
 The program should generate an error message if you accidentally mess up the formatting or enter an invalid value for any of the parameters.
 
-Below, the basics of each simulation test are explained and several examples of the types of physical phenomena that can be simulated are provided.
-If you have further questions about the capabilities and options for any of the simulation tests, please post a question in \href{https://github.com/MikeHeiber/Excimontec/issues}{Issues section} on the Github website.
+The Test Parameters section of the parameters file is used to chose a particular simulation test and set any options specific for each test. In the subsections below, the basics of each simulation test are explained and several examples of the types of physical phenomena that can be simulated are provided.
+If you have further questions about the capabilities and options for any of the simulation tests, please post a question in \href{https://github.com/MikeHeiber/Excimontec/issues}{Issues section} on the GitHub website.
+
+\begin{center}
+\begin{tabular}{ l l }
+\hline
+Test Parameters & \\
+\hline
+N\_tests & \\
+Enable\_exciton\_diffusion\_test & \\
+Enable\_ToF\_test & \\
+ToF\_polaron\_type & \\
+ToF\_initial\_polarons & \\
+Enable\_ToF\_random\_placement & \\
+Enable\_ToF\_energy\_placement & \\
+ToF\_placement\_energy & \\
+ToF\_transient\_start (s) & \\
+ToF\_transient\_end (s) & \\
+ToF\_pnts\_per\_decade & \\
+Enable\_IQE\_test & \\
[…]
