### 2020-08-09 #2307 @MikeHeiber
I've made a significant update to the user manual and paper itself. As before, the latest version can be seen here:
https://github.com/MikeHeiber/Excimontec/tree/joss-paper-update

To the paper, I've added additional information about what KMC simulations can be used for and specifically what types of simulation tests come with v1.0.  I also direct readers interested in more details to the user manual, which provides far more detail than is appropriate for the code paper. I've also specified that TravisCI is doing continuous integration tests instead of simply calling them integration tests.

The user manual has been significantly reorganized and updated to show and explain the parameters within each section of the parameter file and how they related to underlying physical mechanisms that are implemented. I've also added information about the parameter file format and syntax near the beginning. To address concerns with installation instructions, I've added commands to install the GCC compiler and GNU Make.

Clang compiler support has been added and should work.  It's currently being testing and is working on Ubuntu in the TravisCI build matrix. Character encoding issues in the results file will be tackled at a later date.

This should hopefully address all major concerns/issues so that we can move things forward. Thank you to the reviewers for their detailed assessment of everything and helping improve this tool's accessibility with new users.
