### 2020-08-13 MikeHeiber/Excimontec#108 @Luthaf
Updated in the `joss-paper-update` branch!
