### 2020-08-28 #2307 @Luthaf
All of my concerns have been addressed, and I can recommend the publication of this paper in JOSS!

Regarding potential future improvements, I encourage @MikeHeiber to think about modifying the build system for easier integration with different systems (other OS/compilers/CPU architechture/etc.); and continue improving the user manual & documentation to empower both new and power users of his software.
