### 2020-08-28 #2307 @mdavezac
I'm also happy to recommend the publication of the paper in JOSS!

Just a last few comments and recommendation:

- the updated build system did work. However I would recommend adding a few sentences to the installation instructions for non-standard compilers (e.g. run with COMPILER and/or FLAGS set to something) or non-standard compiler names. Such things are not unheard of on HPC platforms. Of course, there is always the option of moving to cmake or similar ;)
- the user manual is pleasingly extensive and informative. It would be even better if it where on the web in html form. Furthermore, it is better at describing what the software does than how to use it in practice. Hence it would be more valuable if the scripts to create the data (and the graphs) therein were more fully available and usable (i.e. not in tar files and with hard-coded paths to external data).
- the input file format is not great, especially considering that parameter sweeps are a likely usage scenario. I expect that the format does not scale well and requires reconciling with each version of the software. I would recommend choosing a standard, flexible, file format, e.g. [TOML](https://github.com/toml-lang/toml), [yaml](https://yaml.org/), or even XML.
