### 2020-08-29 9c9eacd
Changelog Update
- Updated Changelog to note all changes in this PR
- Updated the version string in main to v1.0.0
