### 2020-08-29 35029ed
JOSS Paper Update (#112)
- Resolves #107 
- Resolves #108
- Resolves #109

### Added
- .travis.yml - Build matrix entries and corresponding script code for testing clang compiler
- Changelog - notes about all changes in this PR
- makefile - Detection of versioned g++ compilers
- makefile - Detection of clang compiler and setting the compiler flags
- paper/paper.md - Additional information about what KMC simulations can be used for and specifically what types of simulation tests come with v1.0
- paper/paper.md - Installation instructions for GCC compiler and GNU Make
- README.md - Information a
[…]
