+++ user_manual/references.bib
@@ -186,6 +186,17 @@ doi = "10.21105/joss.01168",
 publisher = "",
 }
 
+@article {jamieson2012cs,
+author = "Fiona C. Jamieson and Ester Buchaca Domingo and Thomas Mc{C}arthy-Ward and Martin Heeney and Natalie Stingelin and James R. Durrant",
+title = "Fullerene crystallisation as a key driver of charge separation in polymer/fullerene bulk heterojunction solar cells",
+year = "2012",
+journal = "Chem. Sci.",
+volume = "3",
+pages = "485-492",
+doi = "10.1039/c1sc00674f",
+publisher = "Royal Society of Chemistry",
+}
+
 @article {kordt2014jctc,
 author = "Pascal Kordt and Ole Stenzel and Bj{\"o}rn Baumeier and Volker Schmidt and Denis Andrienko",
 title = "Parametrization of Extended Gaussian Disorder Models from Microscopic Charge Transport Simulations",
