### 2020-08-29 #2307 @MikeHeiber
Thank you @Luthaf and @mdavezac again for your feedback.  I have filed several issues capturing your final suggestions and will look to implement these additional improvements in the future.

All updates derived from this review process are now merged into master, and I have released v1.0.0 of the software.

@pdebuyl @arfon I have another manuscript in the proofing stage waiting for this process to finish so that I can properly cite this JOSS software paper.  If there is anything that you can do to move this manuscript forward as soon as possible, it would be greatly appreciated.
