+++ paper/paper.bib
@@ -62,8 +62,8 @@ publisher = "",
 author = "Michael C. Heiber",
 title = "Excimontec v1.0: An Open-Source Software Tool for Kinetic {M}onte {C}arlo Simulations of Organic Electronic Devices",
 year = "2020",
-doi = "",
-url = "",
+doi = "10.5281/zenodo.4008122",
+url = "https://doi.org/10.5281/zenodo.4008122",
 }
 
 @article {zuo2019aem,
