### 2020-08-30 #2307 @arfon
> @pdebuyl @arfon I have another manuscript in the proofing stage waiting for this process to finish so that I can properly cite this JOSS software paper. If there is anything that you can do to move this manuscript forward as soon as possible, it would be greatly appreciated.

We need @pdebuyl to make his last review before handing this off to the EiC on rotation. It sounds like this should happen in the next few days.
