+++ paper/paper.md
@@ -24,7 +24,7 @@ bibliography: paper.bib
 
 For over three decades, kinetic Monte Carlo (KMC) simulations have been a powerful computational tool to help understand and optimize organic semiconductor devices, especially photovoltaics, light-emitting diodes, transistors, and thermoelectrics.[@baranovskii2014pssb; @groves2017rpp; @heiber2019chapter; @zuo2019aem]
 KMC simulations use mechanistic models for how excitons and polarons are created, migrate through, and are then eventually removed from the semiconductor layer of a device and can capture the complex interactions between performance and spatial structure that is often not possible using continuum drift-diffusion models. 
-This can then be used to probe a wide variety of phenomena in organic electronics devices, including exciton diffusion and quenching, charge transport, and charge recombination at the full device scale while retaining details regarding nanoscale inhomogeneities. 
+This can then be used to probe a wide variety of phenomena in organic electronic devices, including exciton diffusion and quenching, charge transport, and charge recombination at the full device scale while retaining details regarding nanoscale inhomogeneities. 
 Despite the clear utility of the method, no widespread or standardized software tools have taken hold in the community. 
 Instead, many research groups around the world have maintained private codebases of varying complexity, efficiency, and reliability. 
 As a result, there have been large barriers to entry for new researchers and a lot of repeated effort throughout the community that would have been much better off applied to pushing the capabilities of the technique and further refining the physical models.
