### 2020-09-01 #2307 @MikeHeiber
@pdebuyl The latest version of the paper is on the joss-proofing branch.  I am finished with my edits, so if you don't have any other required edits, you can create the final publication using this branch.  I will then merge it into master once the paper it published.
