### 2020-09-01 #2307 @pdebuyl
@MikeHeiber this is a good time to merge if you wish. Else, the process carries on with the other branch. One of the editors-in-chief will take over the rest of the process.
