### 2020-09-02 #2307 @kthyng
Hi @pdebuyl I saw this but thought I'd give @MikeHeiber a chance to do the merging before I moved ahead. Should I just go ahead now?
