### 2020-09-02 #2307 @pdebuyl
Hi @kthyng thanks for the reply. The merge of the custom branch should be a "fast-forward" merge only and the software release won't change. Please go ahead with the custom branch then, @MikeHeiber is waiting to cite this paper in others, so it is better to proceed.
