### 2020-09-03 #2307 @MikeHeiber
@kthyng Please go ahead and publish the paper in it's current form.  I'll merge the joss-proofing branch in afterwards and also update the branch to include the reference information.
