### 2020-09-03 #2307 @kthyng
Hi @MikeHeiber! I'll help finish up the details for your publication. Please change the metadata for your zenodo archive to exactly match your JOSS paper (title and authors).
