### 2020-09-03 #2307 @kthyng
I just read through your paper @MikeHeiber. It looks fine, except for some reason all the references were listed after the period of the sentence they were used in. I changed this in [this PR](https://github.com/MikeHeiber/Excimontec/pull/117). Please look through and see if you want to merge.
