### 2020-09-04 #2307 @MikeHeiber
@kthyng The Zenodo archive is now updated as requested, and your PR is merged into the joss-proofing branch.
